requirements.txt
pandas==2.1.0
numpy==1.26.4
requests
redis
prettytable
pymysql
chardet

pyecharts==2.0.8
dbutils
apscheduler
tenacity
flask
flask-login
gevent
tornado
sqlalchemy
lark-oapi

mytt
matplotlib
tqdm
ipywidgets
pinyin
pyarmor
setuptools

futu-api
ccxt
baostock
polygon-api-client
alpaca-py
ib_insync
akshare
tqsdk
openai
playwright

pyfolio-reloaded
dtaidistanceczsc
